requirements.txt
streamlit==1.37.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
PyYAML>=6.0
abc-analysis
